[Silver III] Title: 화살표 그리기, Time: 540 ms, Memory: 50496 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/15975. 화살표 그리기/화살표 그리기.java
@@ -12,7 +12,7 @@ public class Main {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		int N = Integer.parseInt(br.readLine());
 		
-		int length = 0;
+		long length = 0;
 		List<ArrayList<Integer>> color = new ArrayList<>();
 		for(int i = 0; i < N; i++) {
 			color.add(new ArrayList<>());
